[AccentColor] Resolve and pass down if Canvas is within installed WebApp scope.
To ensure that the `AccentColor` system keyword is correctly applied
only within the scope of installed WebApps, it's necessary to
introduce a Boolean flag within the
`LayoutTheme::GetAccentColorOrDefault()`. This flag will enable this
method to know whether the content falls under the WebApp's domain,
as it doesn't inherently have access to the `Document`.

Here is part 1 (CL:5785551) and part 2 (CL:5789206) pre-requisites.
This CL is part 3 which will facilitate the passing of this Boolean
value up to the `StyleColor::ResolveColorValue()` function. Subsequent
updates will then integrate this value within the
'LayoutTheme::GetAccentColorOrDefault()` to achieve the
desired behavior.

Bug: 40229450
Change-Id: Ibd421533e2f7fc2bb1c3d9be1e191ab3d512dd41
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5793044
Reviewed-by: Sam Davis Omekara <samomekarajr@microsoft.com>
Commit-Queue: Priya Palanisamy <priyapal@microsoft.com>
Reviewed-by: Joey Arhar <jarhar@chromium.org>
Reviewed-by: Alison Maher <almaher@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1343146}
NOKEYCHECK=True
GitOrigin-RevId: a84d1a78a3c95c5aebf5d81d5fa622e071c9d51b