GPUQueue: check intermediate copy buffer allocation failure
In GPUQueue::CopyFromCanvasSourceImage a wgpu::Buffer that's
mappedAtCreation is created. These buffers may fail creation and when
using dawn::wire when OOM happens or when the
DawnTransferBufferMemoryService decides that the allocation is too big.
Check for allocation failure and return an error when it happens.

Bug: 356077963
Change-Id: Ife03a123835d0182f10a05dc0b7477cb7ccb1319
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5741239
Reviewed-by: Brandon Jones <bajones@chromium.org>
Commit-Queue: Corentin Wallez <cwallez@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1343392}
NOKEYCHECK=True
GitOrigin-RevId: 6cb2176f96f60e90492d1bf8364cf5203652dde4