Apply size containment to ::scroll-marker-group.
There was code for this, but it was never executed, and it was wrong (if
executed, DCHECKs would fail).

It was never executed, because ApplyBaseStyleNoCache() is called for the
originating element, not the pseudo element. Additionally,
ScrollMarkerGroupUserAgentDeclaration() would set the `content` property
to a simple keyword, but it expects a list. See
Contain::ParseSingleValue().

Two tests had to be updated, so that they set a fixed block-size on the
scroll marker group, since the children are not supposed to affect its
size (which is what the tests used to assume).

scroll-marker-002.html actually relied on two layout passes in order to
render as expected, since it places the scroll marker group before the
scroller (and the scroll marker group is populated after the scroller
has been laid out, which will not replace the scroll marker group
fragment itself, only its children (the markers)). We *would* typically
get two layout passes though, since currently we assume that the
viewport gets a vertical scrollbar, and, if that turns out not to be the
case (like in this case), we'll lay out again without the scrollbar. Not
a safe thing to rely on.

Add an additional test that doesn't set an explicit block-size on the
scroll marker group, to make sure that a tall child doesn't affect its
size.

Bug: 332396355
Change-Id: I2ace302c872ff25aaa352b44d688b4c97a7fbb24
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5794890
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Reviewed-by: Daniil Sakhapov <sakhapov@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1343478}
NOKEYCHECK=True
GitOrigin-RevId: 99e15496e6c581d55522b0fad394d1ca109e3f69