Image: Don't GC lazyload-deferred images with decode requests
Before this CL, Blink would allow lazyload-deferred images that were
disconnected from the DOM to be garbage collected. This is because they
didn't technically have a load in progress, and therefore it wasn't
observable to the web whether they were kept around in memory or not.

However, `blink::ScriptPromiseResolver::Dispose()` does not actually
permit disposal if the pending promise associated with it has not been
resolved properly. This caused a crash specifically for
lazyload-deferred images that were disconnected from the DOM (and
therefore garbage collectable), *but* had an unresolved `decode()`
promise out somewhere. When the image was garbage collected,
the ScriptPromiseResolver associated with the underlying ImageLoader
would crash upon disposal.

This CL fixes this problem by amending the
`ImageLoader::HasPendingEvent()` logic to extend the lifetime of the
image if there are pending decode requests. See
https://issues.chromium.org/issues/40071424#comment15 for an analysis.

R=jbroman@chromium.org

Bug: 40071424
Change-Id: Ib2db546ffd578cd5e71e0964a40d41dfe69b17dc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5792476
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Commit-Queue: Dominic Farolino <dom@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1343698}
NOKEYCHECK=True
GitOrigin-RevId: 2d23c2728a17e790cb5afb1894598b16468f277e