FaceGaze: add scroll action
This change implements the scroll action for FaceGaze. In summary, the
user can perform a gesture to enter/exit scroll mode. When scroll mode
is active, moving your head will scroll the container under the mouse
location. We use the center of the screen as a pivot point and
scroll based on the atan2 [1] angle created by the user's head
position. This leads to an intuitive scrolling experience (e.g. head
movements to the left translates to a scroll to the left). To
implement this behavior, a few things needed to happen:

1. A new MouseScroll macro was added
2. A class called ScrollModeController was added to encapsulate all
scrolling logic
3. An extension API is used to communicate the scroll target and
the scroll direction to the browser
4. The browser sends the scroll event to the event sink

[1] https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Math/atan2

Fixed: 355659667
Change-Id: I2fd6f11d4737db2bf3fc76c367633db1b1d74f6f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5769852
Reviewed-by: Kelvin Jiang <kelvinjiang@chromium.org>
Commit-Queue: Akihiro Ota <akihiroota@chromium.org>
Reviewed-by: Amanda Lin Dietz <aldietz@google.com>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1343721}
NOKEYCHECK=True
GitOrigin-RevId: e90912df8492480443baf7024167e7adb2d0b719