Make dedicated & shared workers host know the renderer-side origin
Currently the origin used on the renderer side of dedicated & shared
workers are calculated within the renderer process. Because of this,
the browser side worker host might not be able to know the origin on
the renderer side in case of data: URL workers, which uses a new
opaque origin. This is blocking the calculation of blob: URL origin
on the browser process, since that requires knowing the origin of
the context that creates the blob: URL.

This CL makes it so that the browser side worker host always knows
what the renderer-side origin is, by either:
- Calculating the origin on the browser process and passing it to
  be used on the renderer side. This is done for dedicated workers
  when PlzDedicatedWorker is enabled, and shared workers.
- Passing the origin calculated on the renderer side to the browser
  side, so that the host can save it and refer to it when needed.

Note that the calculated origin on the browser side isn't used by
anything yet and is for informational purposes only for now, but it
will soon be used to resolve Blob URL origins at crrev.com/c/5788808.

This is mostly what is described in awillia@'s doc about this, except
that the StorageKey isn't affected at all:
https://docs.google.com/document/d/1JYSrEy51BgWsTIZQMS_XGG2Z8uFvZi3jvKBkDrSWE0c/edit

Bug: 40051700
Change-Id: I826d8c58709b6c4c6392ed0c05a55d7038c3c7cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5788758
Reviewed-by: Andrew Williams <awillia@chromium.org>
Commit-Queue: Rakina Zata Amni <rakina@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1343751}
NOKEYCHECK=True
GitOrigin-RevId: 415a9006b4facd16ae0e0209fa94ccb970b62a78