Update dir=auto traversal tests to not apply element-only rules to ancestors
This fixes web platform tests to match half of the spec update in
https://github.com/whatwg/html/issues/10488 and
https://github.com/whatwg/html/pull/10561 .  Prior to this change (in
spec and Gecko, although not in the Chromium code) the special rules for
dir=auto traversals that were intended to apply only to dir=auto on an
element and not to dir=auto on its ancestors (relating to
auto-directionality form-associated elements and to slots) were being
applied in one ancestor case, where the special element was an assigned
node of a slot.  This makes the test changes corresponding to the fix to
the HTML spec.

Bug: 576815, 352191350
Change-Id: I6ffa8e22acc6f71777df1165655eaa1c8891867e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5798082
Commit-Queue: David Baron <dbaron@chromium.org>
Reviewed-by: Di Zhang <dizhangg@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1343842}
NOKEYCHECK=True
GitOrigin-RevId: 50a3ec2d5aeaf9b74569e8adc8dbee9f3e35b3e5