Remove references to android_toolchain_canary
This canary build of the toolchain was created for the purposes of
enabling RISC-V builds for initial testing. Now that the NDK r27
toolchain supports building with RISC-V and we no longer use the canary
toolchain, remove it to reclaim the space.

Test: CQ
Bug: 360867127
Change-Id: I1f15b1e878ba158339f7d0454e32af9d8417ac03
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5798143
Reviewed-by: Nico Weber <thakis@chromium.org>
Commit-Queue: Hans Wennborg <hans@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1344018}
NOKEYCHECK=True
GitOrigin-RevId: 44f23c5a2e8c1f437b8219d07c75429985ca60b8