Fix dir=auto traversal to not apply element-only rules to ancestors
This fixes Chromium code to match half of the spec update in
https://github.com/whatwg/html/issues/10488 and
https://github.com/whatwg/html/pull/10561 .  I previously thought the
code was fine, but I realize a small update is needed.  In particular,
the code could apply element-only rules for dir=auto to elements (in
particular, button-type inputs) that are auto-directionality
form-associated elements but were not in the list of exclusions.  This
also fixes the list of exclusions to exclude specifically textarea, so
that any actual text content of input elements (but not the value) is
traversed as required by the spec.

This also removes the no-longer needed is_deferred argument to
ResolveAutoDirectionality.

Bug: 576815, 352191350
Change-Id: I7c13b9661d23f8904727c8298a8dca6360d6d732
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5798880
Commit-Queue: David Baron <dbaron@chromium.org>
Reviewed-by: Di Zhang <dizhangg@chromium.org>
Commit-Queue: Di Zhang <dizhangg@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1344402}
NOKEYCHECK=True
GitOrigin-RevId: 482ca309fca082472da7440fbbed99e8368cb6c0