Add WebDXFeature use counters for device APIs
WebDXFeature::kDeviceOrientationEvents is recorded when a devicemotion, deviceorientation, or deviceorientationabsolute event
listener is registered in a secure context.
https://github.com/web-platform-dx/web-features/blob/main/features/device-orientation-events.yml

WebDXFeature::kDRAFT_Gamepad is recorded when Gamepad.buttons is accessed.
https://github.com/web-platform-dx/web-features/blob/main/features/draft/spec/gamepad.yml

WebDXFeature::kScreenWakeLock is recorded when
navigator.wakelock.request() is called to request wake lock.
https://github.com/web-platform-dx/web-features/blob/main/features/screen-wake-lock.yml

WebDXFeature::kWebBluetooth is recorded when
BluetoothRemoteGATTServer.connect() is called to connect to a
Bluetooth device.
https://github.com/web-platform-dx/web-features/blob/main/features/web-bluetooth.yml

WebDXFeature::kWebNfc is recorded when NDEFReader.scan() is called to
read an NFC tag.
https://github.com/web-platform-dx/web-features/blob/main/features/web-nfc.yml

WebDXFeature::kDRAFT_Serial is recorded when SerialPort.open() is
called to connect to a serial port.
https://github.com/web-platform-dx/web-features/blob/main/features/draft/spec/serial.yml

modified to exclude superficial API usage when no devices are
connected. In both cases, the use counter previously counted the
getDevices() method. Instead, update the use counter when attempting
to open a device with USBDevice.open() or HIDDevice.open().

WebDXFeature: :kWebusb and WebDXFeature::kWebhid use counters are
Change-Id: I3abadd8c32528b5e769f4ff3f36251879c8843e3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5792338
Commit-Queue: Matt Reynolds <mattreynolds@chromium.org>
Reviewed-by: John Delaney <johnidel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1344409}
NOKEYCHECK=True
GitOrigin-RevId: d9a099c82864225a745648288e1d1d2f3e1f3e54