[AccentColor] Enable AccentColor system keyword for installed WebApps
This change enables AccentColor system keyword to use
OS-defined accent-color for installed WebApps ONLY. This uses
WebAppScope parameter in WebPreferences to identify if the content is
for installled webapp and enables AccentColor system keyword in CSS to
access platform defined accent-color.

AccentColor system keyword is not enabled for non-webapp content due to
finger-printing risk. You can refer[1] this CSSWG issue for more detail
with regards to this risk. But based on in-person discussions,
Edge/Chrome decided to enable AccentColor keyword just for installed
Web sites were fingerprinting is not as large of a risk.

[NOTE]: All these changes are behind "CssSystemAccentColor" runtime
feature flag. And this flag is currently in "Experiment"

[1]: https://github.com/w3c/csswg-drafts/issues/10372

Bug: 40229450
Change-Id: Idf3c6f6ce50fc739b56fcc19036e2a9ee217cde8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5766135
Reviewed-by: Alison Maher <almaher@microsoft.com>
Commit-Queue: Priya Palanisamy <priyapal@microsoft.com>
Reviewed-by: Sam Davis Omekara <samomekarajr@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1344525}
NOKEYCHECK=True
GitOrigin-RevId: 5da8314cb77ca818d91111de8f33efa325a62315