Unship KeyboardFocusableScrollers, temporarily
See [1] for details, but due to compat issues, we are going to
disable the feature while we make some adjustments to the
behavior.

This was enabled by default in [2], which landed in M129, so this
CL will need to be merged back to 129.

[1] https://issues.chromium.org/issues/361072782#comment7
[2] https://chromium-review.googlesource.com/c/chromium/src/+/5775613

Fixed: 361072782
Bug: 337158524
Change-Id: I23b451f3ec7db2f8669c62b0cb69c1df6d84edeb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5801654
Auto-Submit: Mason Freed <masonf@chromium.org>
Reviewed-by: Di Zhang <dizhangg@chromium.org>
Commit-Queue: Di Zhang <dizhangg@chromium.org>
Commit-Queue: Mason Freed <masonf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1344534}
NOKEYCHECK=True
GitOrigin-RevId: 9d1ddc1570946d5353a544a6a65bfdf6bd1f751b