Remove tokens_ from TokenSequence.
Nothing cares about the entire list of tokens anymore, so we can do
with (mostly) the first and last tokens. The first token is really
only ever used for fallback sequences, too, but at least this is
a strong reduction in what we need to store.

Change-Id: Ib26e70a9fa27fbb412df9a5d32a65754132ea585
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5797270
Commit-Queue: Steinar H Gunderson <sesse@chromium.org>
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1344633}
NOKEYCHECK=True
GitOrigin-RevId: 8e710e20168d9ac746c6dc932020672b63a16be1