Snav: Removing Unnecessary Media Occlusion Checks
This code was applied when it was not possible to determine media occlusion using a hit test.

Currently, we can use a hit test to identify the media controls that are covering the media element.
By traversing from there, we can find the media element and determine its visibility.

Therefore, the code that previously set |IsMediaElement| to true without checking for occlusion can be removed.

Bug: 41454746
Change-Id: I5733579dbf49a3c78f531b4534c07101e025f7d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5796428
Reviewed-by: JunHo Seo <junho0924.seo@lge.com>
Commit-Queue: sundoo kim <sundoo.kim@lge.com>
Reviewed-by: Stefan Zager <szager@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1344687}
NOKEYCHECK=True
GitOrigin-RevId: d06eb954918235db7e92c78493fdc8b1cf9b2729