Invalidation tracing: Fix incorrect lookup after invalidation set delete
InvalidationSetToSelectorMap uses InvalidationSet pointers as lookup
keys but does not AddRef them. This is good in that it avoids behavior
changing side effects when the map is active during tracing. However,
it also means an InvalidationSet could get deleted and the memory
subsequently reused for a new InvalidationSet, which would lead to
incorrect results on later lookups.

This CL adds a notification to InvalidationSetToSelectorMap when an
InvalidationSet is destroyed so that any existing map entries for them
can be deleted.

While developing this fix, I found that tests did not need to call
UpdateAllLifecyclePhasesForTest() during their setup because
PageTestBase::SetBodyInnerHTML() was already making the same call.
This CL also removes those redundant calls.

Bug: 337076014
Change-Id: Ibc4ced79e069413fc5ab58ab804271ac33415644
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5801295
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Kevin Babbitt <kbabbitt@microsoft.com>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1344845}
NOKEYCHECK=True
GitOrigin-RevId: 1afb4f716b629859bf0f7619f52ee0eb3e272959