Enforce generateBid()'s selectedBuyerAndSellerReportingIdRequired return.
Historically, a bid is considered k-anon, and allowed to win the
auction, based on the renderURL of the ad (as well as the bidding logic
URL and buyer domain). If a bid is k-anonymous, but when combined with
the reporting IDs provided to the reportWin and reportResult functions
is not "k-anonymous for reporting", those reporting IDs are omitted from
the reporting functions, but the bid is still allowed to win.

`selectedBuyerAndSellerReportingIdRequired` is a new optional return
value from `generateBid()` that instructs the auction to treat as
non-k-anon any bid for which the bid is not k-anonymous for reporting.
This only applies when `generateBid()` returns a value for the new
reporting ID, `selectedBuyerAndSellerReportingId`.

This allows buyers to convey that a bid should not be allowed to win the
auction if its reporting IDs aren't going to be provided to
`reportWin()`. Because `scoreAd()` receives the value of
`selectedBuyerAndSellerReportingIdRequired` returned by `generateBid()`,
this also allows sellers to convey that only bids that are k-anonymous
for reporting, and thus will have the reporting IDs provided to
`reportResult()`, should be allowed to win.

Change-Id: I8518612b555265b1f25b1d0ea65132575f2c5dee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5789032
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Commit-Queue: Orr Bernstein <orrb@google.com>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1344863}
NOKEYCHECK=True
GitOrigin-RevId: 604429be9c7056c1283b621b5e87014b5b7ac51d