[a11y] Generational cache for (Next|Previous)OnLine
AXNodeObject::(Next|Previous)OnLine is recursive. Prior to this patch
we did not cache the results of the parent checks resulting in
wasteful duplication of effort. The next or previous on line element
can change over the life of an AXObject, but cannot change while the
tree is frozen.To address, a generational cache has been added. Each
call to (Next|Previous)OnLine checks if we have a value for the
current generation, resetting the cached values if stale. Once a fresh
value is computed, it is stored in the cache to prevent unnecessary
re-evaluation.

https://pinpoint-dot-chromeperf.appspot.com/job/114d4b74210000

11% improvement in TotalAccessibilityCleanLayoutLifecycleStages
15% improvement in SerializeLifecycleStage

Change-Id: I077b8cc0f5204248b4d558d547c227198a56cec9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5797643
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Commit-Queue: Kevin Ellis <kevers@chromium.org>
Reviewed-by: Lucas Radaelli <lucasradaelli@google.com>
Cr-Commit-Position: refs/heads/main@{#1345001}
NOKEYCHECK=True
GitOrigin-RevId: 7addef8c560a73698ec44f8467d0ac9c2c52ab82