Improve handling of Link header `rel="preload";as="document"` prefetches
When Chrome encounteres a Link rel=preload header on a prefetched
response, it will attempt to promote that to a prefetch instead.
If the Link header has `rel="preload";as="document"`, it will set a
flag indicating that the prefetch should be for a top-level navigation.
Chrome doesn't correctly support building a request when both of these
conditions are true, so this CL makes it so that the `as=document`
parameter was not specified (matching how the code behaves today).

I added a new test for this but have it disabled like some similar
tests that have flakiness issues. I ran the tests manually, though,
using:

./out/Debug/content_browsertests --gtest_filter="*DISABLED_CrossOriginWithInvalidPreloadAsDocument/*" --gtest_also_run_disabled_tests

Bug: 357325599, 347934841
Change-Id: Ifa1729dbfa04261de8cb4babaadcd74a087ac6e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5769944
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Commit-Queue: Andrew Williams <awillia@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1345125}
NOKEYCHECK=True
GitOrigin-RevId: 436331303b8eb05698cee5c12614f8560edf1d49