Add Infrastructure for Hand tracking permission
Adds a new ContentSetting and PermissionType for the HandTracking
permission. Also adds the basic text, icons, and PermissionContext to
prompt for it, though it is not checked/enforced/prompted for at the
moment as that will be handled in a future change.

Bug: 359418633
Change-Id: If02ee436b2262131de8daf2baf3ea50e5fa762c1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5786325
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Reviewed-by: Evan Stade <estade@chromium.org>
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Reviewed-by: Victor Tan <victortan@chromium.org>
Reviewed-by: Elly FJ <ellyjones@chromium.org>
Reviewed-by: Balazs Engedy <engedy@chromium.org>
Reviewed-by: Florian Jacky <fjacky@chromium.org>
Reviewed-by: Peter Beverloo <peter@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1345562}
NOKEYCHECK=True
GitOrigin-RevId: edef12b6eb52291187bed8d94a9ced7a5f1de691