WebNN: Fixed resetAfter option of gru operator in DirectML backend
This CL fixes the implementation of the resetAfter option of the gru
operator in the DirectML backend. After the modification, the gru
operator in the DirectML backend will have the same behavior as other
backends such as tflite.

Since the conformance tests of gru and gruCell operators are based on
the existing gru test of ONNX CTS, and the default value of
`linear_before_reset` in ONNX is false whereas the default value of
`resetAfter` in WebNN is true, we should explicitly set the resetAfter
parameter to false in the test case when the `linear_before_reset` is
not set by ONNX CTS. This CL also modifies the WPT conformance test of
gru and gruCell.

Bug: 359628066
Change-Id: I70feee5bc67d894b2823cc4f4881b77d4e0edd91
Cq-Include-Trybots: luci.chromium.try:win11-blink-rel,mac14.arm64-blink-rel,mac14-blink-rel, linux-blink-rel
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5789662
Reviewed-by: Rafael Cintron <rafael.cintron@microsoft.com>
Reviewed-by: ningxin hu <ningxin.hu@intel.com>
Commit-Queue: Bin Miao <bin.miao@intel.com>
Cr-Commit-Position: refs/heads/main@{#1345781}
NOKEYCHECK=True
GitOrigin-RevId: e804cad7dbec253b6600bc9f30c96dd2cce73fec