[bindings] All partial use of extended idl attribute [NoAllocDirectCall]
With https://chromium-review.googlesource.com/c/v8/v8/+/5797463, V8 only
uses fast API calls when the number of parameters provided by the
application and the number of parameters expected by the API function
match. With that it is now possible to annotate only some overloads of a
function, as long as all overloads with the same number of parameters
have the `[NoAllocDirectCall]` extended attribute, or none.

This CL therefore relaxes the check in the bindings generator
accordingly.

R=japhet@chromium.org

Bug: 40259610
Change-Id: I45ca558acf4567dac0873b569f4366de287bd2fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5804326
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Reviewed-by: Nate Chapin <japhet@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1345828}
NOKEYCHECK=True
GitOrigin-RevId: 4a495cbe00ff3a7e8f75cce1cdde82a0d04e3a15