Remove LCP mouseover heuristic code.
This code was added as an experiment to disable late LCP entries
caused by a mouseover-to-zoom interaction, but doesn't reliably
prevent LCP in many cases.

Future work will replace this with a heuristic based on attributing new
LCP candidates to JavaScript tasks which can be traced back to a mouse
event listener. In the meantime, removing this code will allow the
existing code to be refactored and simplified.

Bug: 40211421
Change-Id: Ib2bbe39bfb2f9c988becaf7b72e479b312fb3ba4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5789151
Reviewed-by: Michal Mocny <mmocny@chromium.org>
Reviewed-by: Robert Flack <flackr@chromium.org>
Commit-Queue: Ian Clelland <iclelland@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1346021}
NOKEYCHECK=True
GitOrigin-RevId: 7637ecaa727d42f821925bc339a953e550296d05