DeferRenderTasksAfterInput: Fix overlapping use cases
For the DeferRenderTasksAfterInput experiment, if discrete input runs
while in a continuous gesture use case, the discrete input use case
takes precedent (except for touchstart). This WAI, but if we reach the
50ms timeout, the use case is not resetting to the previous one since
there is still time remaining in the gesture. This could extend the
discrete input use case longer than was intended if no frame was
produced.

To fix this, this CL changes exposes `UserModel`'s two time-remaining
methods: `TimeLeftInContinuousUserGesture()` and
`TimeLeftUntilDiscreteInputResponseDeadline()`, and removes the
combined `TimeLeftInUserGesture`. Then `ComputeUseCase()` uses these
individually to determine the use case and time until the policy needs
to be recomputed. Additionally, several unit tests are added and
updated.

Bug: 350540984
Change-Id: I9687e3cbcc618d5be0068bc581ddc9841e82ff47
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5805455
Commit-Queue: Scott Haseley <shaseley@chromium.org>
Reviewed-by: Michal Mocny <mmocny@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1346105}
NOKEYCHECK=True
GitOrigin-RevId: b65c83d0a6d25ec6a52cb858eccbd051abc0b2ad