Only reinitialize the v8::Context if necessary when swapping frames
Re-initializing the v8::Context after navigation is important if a
window was ever scripted, since it ensures that any extant JS references
to the window seamlessly refer to the new v8::Context instead. If a
window was *never* scripted, there cannot be references to the window,
and re-initialization can be skipped.

However, frame swap has always unconditionally called Initialize(). With
RenderDocument, this is problematic since the Java object bridge has
some injection edge cases that are sensitive to the exact timing of the
`DidClearWindowObject()` call. To restore the pre-RenderDocument
behavior, frame swap should similarly skip re-initialization if the
window was never scripted.

Also update the web frame swap tests to have some basic coverage for
this, along with some supporting fixes:
- avoid calling SwapIn() directly and use CommitNavigation() to trigger
  the swap instead, which mirrors how the standard navigation commit
  flow works.
- removing the name="..." attributes from frame-a-b-c.html, since a
  named iframe forces the window proxy to be initialized
- creating named-frame-a-b-c.html for the single frame swap test that
  *does* require named frames

Bug: 40615943
Change-Id: I37391be5f7764a9ac449b1966d7172fbc19ea189
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5801523
Commit-Queue: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1346282}
NOKEYCHECK=True
GitOrigin-RevId: 112acb7642c28465726ada279c6e237a0bea3e64