BASE_FEATURE_PARAM: GetFieldTrialParamByFeatureAsString
Until today, we didn't have a GetFieldTrialParamByFeatureAs<Type>
variant for std::string and use base::FeatureParam<std::string>
instead. But it's really useful to have the std::string variant
because there is a common pattern where we query the std::string
parameter value only once, and parse it to construct a per-process
static singleton-like structure.

This patch defines GetFieldTrialParamByFeatureAsString as we do for
other types, and replace existing two patterns with this base API.

One is kAutoSpeculationRulesConfig that is the pattern we query
only once to construct a static local per-process struct.

The other is miracle parameter usage where they already defined
a similar function. It didn't permit the empty string. But as
changing it to permit empty strings would not affect compatibility,
we will use the same base API here.

Bug: 360291576
Change-Id: Iba7eb926980ffbc7c8299f6242718751de862363
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5797702
Reviewed-by: Minoru Chikamune <chikamune@chromium.org>
Commit-Queue: Takashi Toyoshima <toyoshim@chromium.org>
Reviewed-by: Richard (Torne) Coles <torne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1346563}
NOKEYCHECK=True
GitOrigin-RevId: 267198895a3419aa26f7c8a690b9aa98f0552f9b