Input event data should not be null when dropped data has interchange newline at the end or start
As per spec
https://www.w3.org/TR/input-events-1/#overview, event data for input
events with certain inputType should be set with the data under
operation. Currently we save this data in the |PerformTrivialReplace|
method but if the dropped data has an interchange newline at the start
or end, then the method returns early and this data is not saved. This
change will set the data just after checking if we are dealing with text
node.

Bug: 40737336
Change-Id: Ib52fe4f7e8d8bfa00130894cbea904fe150bcf8b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5782430
Commit-Queue: Utkarsh Pathak <utpathak@microsoft.com>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Reviewed-by: Siye Liu <siliu@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1346573}
NOKEYCHECK=True
GitOrigin-RevId: a55efecc7a94210fb7de6afabaa8a927f2ba2c14