Use a HeapHashMap for UserScrollableAreas
It was noted in a comment in LocalFrameView that we could use a hash
map to speed up lookup (we need to be able to find a scrollable area
matching a given CompositorElementId). In preparation for using this
lookup more frequently in a future CL, we make the change to using a
hash map here.

credit: suggested by wangxianzhu@.

Bug: None
Change-Id: I8679a190e4bdbf80082742374fa9f6f2bcc358ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5792862
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Reviewed-by: Robert Flack <flackr@chromium.org>
Commit-Queue: Ian Vollick <vollick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1346725}
NOKEYCHECK=True
GitOrigin-RevId: db9bcc5daf85a594795cb31e1ea5ca62a6eac8df