Convert most at-rule prelude parsing to the streaming parser.
Instead of consuming the prelude up-front as a block, we now parse it
as any regular part of the stream. (This is hard in some places,
mostly due to interactions with the non-streaming parser; these have been
left for later. Examples include @media and @supports, which have
complex sub-parsers on their own.)

ConsumeEndOfPreludeForAtRuleWithBlock() (and similarly
...WithoutBlock()) now test for the end of the prelude themselves,
so the callers do not need to do that. This works both for the new
streaming prelude parsers and the callers that still call
ConsumeAtRulePrelude(). I've tried to more clearly mark which parts
of the function would be about prelude parsing and which are about
parsing the block (if any); previously, we'd often enter the block
and only then start parsing/error-checking the prelude.

As a side effect, our parsing structure is slightly more regular,
so we return observer signals more often for extra junk in the preludes.
It seems also @font-face would accept junk tokens after its prelude,
which has also been fixed by the increased regularity.

Change-Id: Idbe262bd4e307bd5dba44586355edb5f6ed231c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5803408
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Commit-Queue: Steinar H Gunderson <sesse@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1346759}
NOKEYCHECK=True
GitOrigin-RevId: 79ef491e8ecece78d1db98c2acfb4c79da120fbb