Enable <dialog popover autofocus> new behavior
This pulls out one bit of the DialogNewFocusBehavior that is
specific to popovers, and is per-spec. Since this shouldn't
be a breaking change, this just fixes the bug and adds a
killswitch in case of compat issues. But there shouldn't be
any issues with this.

This fixes the last three popover Interop2024 tests:
https://wpt.fyi/results/html/semantics/popovers/popover-focus.html?label=stable

See this discussion for the more general dialog focus algorithm:
 https://github.com/whatwg/html/issues/9640

Bug: 40369887
Change-Id: I2b197293f83d76ffc1834429e77dadff57b5b587
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5809090
Commit-Queue: Joey Arhar <jarhar@chromium.org>
Auto-Submit: Mason Freed <masonf@chromium.org>
Reviewed-by: Joey Arhar <jarhar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1346786}
NOKEYCHECK=True
GitOrigin-RevId: e3092e4a17901dc6f39a418b8792b5c8320ccb0c