Better control of breaks before trailing decorations.
Add a fragment builder flag to control this. This isn't something that
FinishFragmentation() can determine on its own. Unconditionally allowing
it for tables, like we used to, would result in an infinite loop, if the
table had already decided that we were at the last fragment, and we'd
then add a break before trailing border+padding in the cloning
box-decoration-break model.

Move some comments around, so that they are next to the code they're
relevant for.

Bug: 40415661
Change-Id: I026323731f494595a17ae6437a7454131ec9e7ce
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5811844
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1346904}
NOKEYCHECK=True
GitOrigin-RevId: f7afd40fcbb68a034197b79db1923a4ce515e257