Stop to log differences in UTC timestamps.
In https://chromium-review.googlesource.com/c/chromium/src/+/5314759 and https://chromium-review.googlesource.com/c/chromium/src/+/5660450 we added a log to see if we could see a offset between an UTC clock with static member offset (emulating the WebRTC NTP clock) and the MonotonicTimeToPseudoWallTime used to make UTC times in blink.

We see that there is a difference in a few procent of calls, that was the goal of this and there is no more info to be gain.

We have also learned how to reproduce the issue locally, which make the data even more useless.

Since the histogram was only useful to prove to prove that the issue was possible, I see no reason to mark it as obsolete before removing it.

OBSOLETE_HISTOGRAMS=Not needed anymore.

Bug: 333359951
Change-Id: I80772c16b8b2a094421a5e2e0149fdeed79f68cc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5803967
Reviewed-by: Johannes Kron <kron@chromium.org>
Reviewed-by: Harald Alvestrand <hta@chromium.org>
Commit-Queue: Olov Brändström <brandstrom@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1347308}
NOKEYCHECK=True
GitOrigin-RevId: 60703ec6d9524e35dc25f8133fefa2ebb518fdf2