Replace base-select svg with disclosure-open
This was suggested here:
https://github.com/w3c/csswg-drafts/issues/10380#issuecomment-2189397947

This patch also adds min-inline-size and min-block-size to
::select-fallback-button because changing the icon caused the total size
of the button to decrease which revealed the lack of min sizing which is
already tested by
select-accessibility-minimum-target-size.tentative.html

The UA select:open rules caused a performance regression for large
<select>s because it caused a style recalc immediately after opening the
appearance:auto picker which in re-serializes all of the data for the
popup and sends it to the popup, which is very slow for large selects. I
addressed this by starting to match :open before creating the popup in
MenuListSelectType::ShowPopup.

Bug: 1511354
Change-Id: Id7b3b0aa6c9c79299240a7212023f849f3703fb0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5660759
Commit-Queue: Joey Arhar <jarhar@chromium.org>
Reviewed-by: David Baron <dbaron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1347324}
NOKEYCHECK=True
GitOrigin-RevId: d3040ea658cf8f07522a49242155cdc22f43cf55