Clean up input handling in CSSColorFunctionParser
Consume whitespace and close the scope for the block guard as soon as
we've finished handling tokens. As a consequence, promote the variable
tracking whether the input contained an alpha value out of the block
guard scope and rename it to reflect that it's not an expectation but a
reality.

Bug: 325309578
Change-Id: Ibad65e0bba04309a66ebb728e137477868c5a51c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5813472
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Kevin Babbitt <kbabbitt@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1347401}
NOKEYCHECK=True
GitOrigin-RevId: 7e22df9f72b69d0f85f6dec81c7c463262145f85